Improve error handling in Delta Sharing Server (#32)
Currently Armeria is doing the request parsing work. When a request is invalid, such as invalid json (#27), or incorrect type (`maxResults` is not an int), it will throw an exception and report 500 Internal Error.

In this PR, we catch any unhandled exception when process a request, and convert it to `DeltaInternalException` when processing a request. Then we scope all exceptions that may contain sensitive information to a single type `DeltaInternalException`. Hence, `DeltaSharingServiceExceptionHandler` can safely expose the messages in the following exceptions to the user for better user experience and report 400 Bad Request:
- `scalapb.json4s.JsonFormatException`: Valid json string but invalid field type (#27).
- `com.fasterxml.jackson.databind.JsonMappingException`: Invalid json string.
- `java.lang.NumberFormatException`: `maxResults` is not an int.

Fixes #27