Refresh pre signed urls in Spark connector (#69)
Currently, when reading a Delta Sharing table, we generate the pre signed urls first and use them to read. However, for a long running query, the pre signed urls may expire before we try to read them.

This PR adds a pre signed url cache in Spark driver. It will refresh pre signed urls in a background thread. Tasks running in Spark executors will talk to the Spark driver to fetch pre signed urls. This will ensure when we read a file in an executor, it will always get the latest valid pre signed url.