Add User-Agent request header to Delta Sharing client (#75)
This PR adds the User-Agent request header to Delta Sharing client so that the server can use this information to track client metrics and it's also easier to debug issues. Here are examples of the header:

- Spark connector

> Delta-Sharing-Spark/0.3.0-SNAPSHOT Spark/3.1.1 Hadoop/3.2.0 Mac_OS_X/10.14.6 Java_HotSpot(TM)_64-Bit_Server_VM/25.171-b11 java/1.8.0_171 scala/2.12.10 java_vendor/Oracle_Corporation

- Pandas connector

> Delta-Sharing-Python/0.3.0.dev0 pandas/1.2.4 PyArrow/4.0.0 Python/3.9.5 System/macOS-10.14.6-x86_64-i386-64bit

There is no great way to write unit tests for this. I manually changed the code in server to print the User-Agent header and ran all client tests to verify the content.

Closes #57