Add a limit parameter to 'load_as_pandas' (#76)
This PR adds a new parameter `limit` to `load_as_pandas`. Users can use this parameter to limit the number of rows to be loaded into a pandas DataFrame. It can be used to explore a shared table without loading the entire table into the memory.

This PR also upgrades pyarrow to 4.0.0 so that we can use [dataset.head](https://issues.apache.org/jira/browse/ARROW-9731) to load only the first N rows.

Closes #62