misc: api-description.yml license aligned with delta-sharing
In discussion with original author of
delta-sharing-protocl-api-description.yml file (R. Tyler Croy),
listing this file as AGPL 3.0 was a clerical error when it
was brought over from its original project.
This modification fixes the clerical error by removing the
license information from the file, thus aligning it with the
license of the project. This is being done at the request
of the author.

Here is the exchange with @rtyler that prompted this mod:
"
That's a good catch @Christian Daudt , I originally authored that
file as part of riverbank, a rust delta-sharing server which I
had originally licensed as AGPL before it entered the
delta-incubator.
When I updated to the Apache license I forgot about that little note,
that's a clerical error and can be cleaned up.
"

Signed-off-by: Christian Daudt <christian.daudt@improving.com>