Update PresignedUrlCache exception message for clarity
Why:
Today, when a job doesn't access the delta table before `expireAfterAccessMs`, the presigned url is removed from the cache. When the job does try to access that table, the entry is removed from the concurrent hash map and an IllegalStateException is thrown. This exception message says the table was "removed". The message is confusing if the developer doesn't understand the table was "removed" from the cache. It is not unreasonable for the developer to be confused that the delta table was suddenly removed completely!

How this works:
This changes the exception message to carry the expireMilliseconds information in the stack trace. This should help developers understand and adjust their job configurations