added support for OAuth Client Credentials grant in the Scala client (#553)
**Added OAuth Client Credentials Support in Scala Client**
Overview:
This PR introduces support for OAuth Client Credentials in the Scala client. 

Adds a CredentialProvider trait and two implementations: 
* BearerTokenAuthProvider and OAuthClientCredentialsAuthProvider.

`CredentialProvider` Trait:
Added the CredentialProvider trait which includes a method CredentialProvider#addAuthHeader(httpRequest: HttpRequestBase). This method is responsible for adding the authentication header to the HTTP request.
Implementations of CredentialProvider:

`BearerTokenAuthProvider`: A simple implementation that uses a provided bearer token.

`OAuthClientCredentialsAuthProvider`: This implementation handles OAuth client credential grants. It manages the access token internally and refreshes it when it is about to expire (within the next 10 minutes), using the clientId and clientSecret to obtain a new token.
Functionality:

The DeltaShringClient now invokes credentialProvider.addAuthHeader() before making any requests, ensuring that the appropriate authentication header is included.
OAuth Token Management:

The OAuthClientCredentialsAuthProvider efficiently manages the OAuth access token. It only exchanges the clientId and clientSecret for a new access token when the current token is nearing expiration

New format for the profile share file:
```
{
    "shareCredentialsVersion": 2,
    "type": "persistent_oauth2.0",
    "endpoint": "https://example.com/delta-sharing/",
    "tokenEndpoint": "https://example.com/oauth/v1/token",
    "clientId": "ABCE",
    "clientSecret": "XYZE",
    "scope" : "abc" // optional parameter
}

```