Python OAuth client_credentials support (#563)
The current OAuth implementation in the Python client does not reuse access tokens. Instead, it exchanges the client-id and client-secret for a new access token with every request. This behavior increases the load on the tokenEndpoint and introduces unnecessary latency in the request processing.

This PR updates the OAuth implementation to oauth reuse access tokens, aligning the Python client with the OAuth behavior of the Scala client, as detailed in this [PR for the Scala client](https://github.com/delta-io/delta-sharing/pull/553).

For additional context on the usage of auth_provider and its functionality, please refer to the description in the original PR for the Spark client.